+++ src/widgets/file_notes.py
@@ -1,58 +1,15 @@
 from loguru import logger
 
 from PyQt6.QtCore import Qt, QDateTime, pyqtSlot
-from PyQt6.QtGui import QMouseEvent, QFocusEvent
-from PyQt6.QtWidgets import (QWidget, QTextEdit, QSizePolicy,
-    QMessageBox, QVBoxLayout, QScrollArea, QAbstractScrollArea,
-    QMenu,
+from PyQt6.QtGui import QMouseEvent
+from PyQt6.QtWidgets import (QWidget, QSizePolicy, QMessageBox,
+    QVBoxLayout, QScrollArea, QAbstractScrollArea, QMenu,
 )
 
 from ..core import app_globals as ag, db_ut, utils
 from .file_note import fileNote
+from .note_editor import noteEditor
 
-
-class noteEditor(QTextEdit):
-    def __init__(self, parent = None) -> None:
-        super().__init__(parent)
-        self.note: fileNote = None
-        self.branch = None
-        self.setStyleSheet(ag.dyn_qss['note_editor'][0])
-        self.setFocusPolicy(Qt.FocusPolicy.StrongFocus)
-
-    def focusOutEvent(self, e: QFocusEvent):
-        logger.info(f'{e.lostFocus()}')
-        if e.lostFocus():
-            ag.signals_.user_signal.emit('SaveEditState')
-        super().focusOutEvent(e)
-
-    def start_edit(self, note: fileNote):
-        self.note = note
-        self.setText(db_ut.get_note(
-            self.get_file_id(), self.get_note_id()
-            )
-        )
-
-    def set_branch(self, branch):
-        self.branch = branch
-
-    def get_file_id(self) -> int:
-        return self.note.get_file_id() if self.note else 0
-
-    def get_note_id(self) -> int:
-        return self.note.get_note_id() if self.note else 0
-
-    def get_branch(self) -> str:
-        return self.branch
-
-    def get_text(self):
-        return self.toPlainText()
-
-    def get_note(self) -> fileNote:
-        return self.note
-
-    def set_note(self, note: fileNote):
-        self.note = note
-        logger.info(f'{self.note.get_file_id()=}, {self.note.get_note_id()=}')
 
 class notesContainer(QScrollArea):
     def __init__(self, editor: noteEditor, parent: QWidget=None) -> None:
@@ -60,6 +17,7 @@ class notesContainer(QScrollArea):
 
         self.editor = editor
         self.editing = False
+        self.ini = True
         self.set_ui()
 
         self.file_id = 0
@@ -82,6 +40,8 @@ class notesContainer(QScrollArea):
         self.scrollWidget.setObjectName("scrollWidget")
         self.setWidget(self.scrollWidget)
         self.scroll_layout = QVBoxLayout(self.scrollWidget)
+        self.scroll_layout.setContentsMargins(9,0,6,0)
+        self.scroll_layout.setSpacing(6)
         self.scroll_layout.setObjectName('scroll_layout')
         self.setStyleSheet("border: none;")
 
@@ -97,9 +57,10 @@ class notesContainer(QScrollArea):
 
     def go_action(self, act_text: str):
         file_id = self.editor.get_file_id()
-        branch = self.editor.get_branch()
+        branch = ','.join((str(i) for i in self.editor.get_branch()))
+        # logger.info(f'{branch=}')
         ag.signals_.user_signal.emit(
-            f"file-note: {act_text}/{file_id}-{branch}"
+            f"file-note: {act_text}\\{file_id}-{branch}"
         )
 
     def is_editing(self):
@@ -110,7 +71,6 @@ class notesContainer(QScrollArea):
         self.edited_file_in_status(state)
 
     def edited_file_in_status(self, show: bool):
-        logger.info(f'{show=}')
         if show:
             file_id = self.editor.get_file_id()
             filename = db_ut.get_file_name(file_id)
@@ -126,13 +86,19 @@ class notesContainer(QScrollArea):
         self.set_notes_data()
 
     def set_notes_data(self):
+        self.setUpdatesEnabled(False)
         self.clear_layout()
         self.scroll_layout.addStretch(1)
+        if self.ini:
+            self.ini = False
+            self.setUpdatesEnabled(True)
+            return
         data = db_ut.get_file_notes(self.file_id)
         for row in data:
-            note = fileNote(*row[1:])
+            note = fileNote(*row[1:], self.file_id)
             note.set_text(row[0])
             self.add_item(note)
+        self.setUpdatesEnabled(True)
 
     def clear_layout(self):
         for i in reversed(range(self.scroll_layout.count())):
@@ -153,30 +119,29 @@ class notesContainer(QScrollArea):
 
     def update_note(self):
         note: fileNote = self.editor.get_note()
-        file_id = note.get_file_id()
+        note_file_id = note.get_note_file_id()
         note_id = note.get_note_id()
         txt = self.editor.get_text()
 
         if note_id:
-            self.scroll_layout.removeWidget(note)
-            ts = db_ut.update_note(file_id, note_id, txt)
+            ts = db_ut.update_note(note_file_id, note_id, txt)
         else:
-            ts, note_id = db_ut.insert_note(file_id, txt)
-            note.set_note_id(note_id)
-            note.set_creation_date(ts)
+            ts = db_ut.insert_note(note_file_id, txt)
 
-        note.set_modification_date(ts)
+        file_id = note.get_file_id()
+        ag.add_history_file(file_id)
+
         if self.file_id == file_id:
-            note.set_note_text(txt)
-            self.add_item(note)
             self.update_date_in_file_list(ts)
+            self.set_notes_data()
 
     def update_date_in_file_list(self, ts: int):
         if ts > 0:
-            a = QDateTime()
-            a.setSecsSinceEpoch(ts)
+            last_note_date = QDateTime()
+            last_note_date.setSecsSinceEpoch(ts)
             ag.file_list.model().update_field_by_name(
-                a, "Date of last note", ag.file_list.currentIndex()
+                last_note_date, "Date of last note",
+                ag.file_list.currentIndex()
             )
 
     @pyqtSlot(fileNote)
@@ -198,9 +163,9 @@ class notesContainer(QScrollArea):
         if utils.show_message_box(
             'delete file note',
             'confirm deletion of note',
-            QMessageBox.StandardButton.Ok |
+            btn=QMessageBox.StandardButton.Ok |
             QMessageBox.StandardButton.Cancel,
-            QMessageBox.Icon.Question
+            icon=QMessageBox.Icon.Question
         ) == QMessageBox.StandardButton.Ok:
             self.scroll_layout.removeWidget(note)
             note.deleteLater()