+++ src/widgets/find_files.py
@@ -80,7 +80,7 @@ class findFile(QWidget):
 
         if db_ut.exists_file_with_name(name, case, word):
             ag.signals_.user_signal.emit(
-                f'find_files_by_name/{name},{int(case)},{int(word)}'
+                f'find_files_by_name\\{name},{int(case)},{int(word)}'
             )
             ag.save_settings(SEARCH_FILE=(name, case, word))
             self.close()