+++ src/widgets/open_db.py
@@ -5,7 +5,6 @@ from PyQt6.QtCore import Qt, pyqtSlot, QPoint
 from PyQt6.QtGui import QKeySequence, QShortcut
 from PyQt6.QtWidgets import (QFileDialog, QLabel,
     QListWidgetItem, QVBoxLayout, QWidget, QMenu,
-    QApplication,
 )
 
 from ..core import create_db, icons, utils, app_globals as ag
@@ -43,11 +42,9 @@ class OpenDB(QWidget, Ui_openDB):
 
         self.setupUi(self)
         self.msg = ''
+        self.path = ''
 
         self.restore_db_list()
-
-        self.open_btn.setIcon(icons.get_other_icon("open_db"))
-        self.open_btn.clicked.connect(self.add_db)
 
         self.listDB.doubleClicked.connect(self.item_click)
         self.listDB.setContextMenuPolicy(Qt.ContextMenuPolicy.CustomContextMenu)
@@ -55,26 +52,19 @@ class OpenDB(QWidget, Ui_openDB):
         self.listDB.currentItemChanged.connect(self.row_changed)
         self.listDB.setCurrentRow(0)
 
-        self.input_path.textEdited.connect(self.style_input_path)
-        self.input_path.editingFinished.connect(self.finish_edit)
-        self.input_path.setContextMenuPolicy(Qt.ContextMenuPolicy.CustomContextMenu)
-        self.input_path.customContextMenuRequested.connect(self.path_menu)
-
         escape = QShortcut(QKeySequence(Qt.Key.Key_Escape), self)
         escape.activated.connect(self.close)
-        self.set_tool_tip()
 
     @pyqtSlot(QListWidgetItem, QListWidgetItem)
     def row_changed(self, curr: QListWidgetItem, prev: QListWidgetItem):
         wid: listItem = self.listDB.itemWidget(curr)
-        self.input_path.setText(wid.get_db_name())
-        self.set_tool_tip()
+        self.path = wid.get_db_name()
 
     @pyqtSlot(QPoint)
     def item_menu(self, pos: QPoint):
         item = self.listDB.itemAt(pos)
         if item:
-            db_name = self.input_path.text()
+            db_name = self.path
             menu = self.db_list_menu(
                 db_name[db_name.rfind('/')+1:]
             )
@@ -99,30 +89,6 @@ class OpenDB(QWidget, Ui_openDB):
         menu.addAction(f'Delete DB "{db_name}" from list')
         return menu
 
-    def set_tool_tip(self):
-        self.input_path.setToolTip(
-            'Enter path to create database or choose from '
-            'the list below. Esc - to close without choice'
-        )
-        self.input_path.setPlaceholderText(
-            "Enter path to open/create database. "
-            "Esc - to close without choice"
-        )
-
-    @pyqtSlot(QPoint)
-    def path_menu(self, pos: QPoint):
-        menu = QMenu(self)
-        menu.addAction("Copy message")
-        action = menu.exec(self.input_path.mapToGlobal(pos))
-        if action:
-            self.copy_message()
-
-    def copy_message(self):
-        if self.input_path.text():
-            QApplication.clipboard().setText(self.input_path.text())
-        else:
-            QApplication.clipboard().setText(self.input_path.placeholderText())
-
     def restore_db_list(self):
         db_list = utils.get_app_setting("DB_List", []) or []
         for it in db_list:
@@ -141,30 +107,17 @@ class OpenDB(QWidget, Ui_openDB):
     def remove_item(self, item: 'QListWidgetItem'):
         self.listDB.takeItem(self.listDB.row(item))
 
-    def style_input_path(self, text: str):
-        self.input_path.setStyleSheet(ag.dyn_qss['input_path_edited'][0])
-        self.input_path.setToolTip('Esc - to close without choice')
-
-    def finish_edit(self):
-        if self.msg:
-            return
-        db_name = self.input_path.text()
-        if db_name:
-            self.add_db_name(Path(db_name).as_posix())
-
     def show_error_message(self):
         if not self.msg:
             return
-        self.input_path.setStyleSheet(ag.dyn_qss['input_path_message'][0])
+        ag.app.ui.msg.setStyleSheet(ag.dyn_qss['input_path_message'][0])
+        ag.app.ui.msg.setText(self.msg)
 
-        self.input_path.clear()
-        self.input_path.setPlaceholderText(self.msg)
-        self.input_path.setToolTip(self.msg)
+        ag.app.ui.msg.setToolTip(self.msg)
         self.msg = ''
 
     def add_db_name(self, db_name:str):
         db_ = db_name.strip()
-        logger.info(db_)
         if self.open_if_here(db_):
             return
 
@@ -172,7 +125,6 @@ class OpenDB(QWidget, Ui_openDB):
 
     def open_if_ok(self, db_name: str):
         if self.verify_db_file(db_name):
-            logger.info(f"{self.msg=}")
             self.add_item_widget(db_name)
             self.open_db(db_name)
             return
@@ -203,9 +155,7 @@ class OpenDB(QWidget, Ui_openDB):
                 False otherwise
         """
         file_ = Path(file_name).resolve(strict=False)
-        logger.info(self.input_path.text())
-        logger.info(self.input_path.placeholderText())
-        self.input_path.setText(file_name)
+        self.path = file_name
         if file_.exists():
             if file_.is_file():
                 if create_db.check_app_schema(file_name):
@@ -229,14 +179,14 @@ class OpenDB(QWidget, Ui_openDB):
 
     @pyqtSlot()
     def item_click(self):
-        self.open_db(self.input_path.text())
+        self.open_db(self.path)
 
     def open_db(self, db_name: str):
         ag.signals_.get_db_name.emit(db_name)
         self.close()
 
     def open_in_new_window(self, db_name: str):
-        ag.signals_.user_signal.emit(f'Setup New window/{db_name}')
+        ag.signals_.user_signal.emit(f'Setup New window\\{db_name}')
         self.close()
 
     def get_item_list(self) -> list:
@@ -249,7 +199,5 @@ class OpenDB(QWidget, Ui_openDB):
         return items
 
     def close(self) -> bool:
-        self.msg = 'closing'
-        logger.info(f"{self.msg=}")
         utils.save_app_setting(DB_List=self.get_item_list())
         return super().close()